requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
gunicorn